TASK: Ensure before root node update all workspaces are empty
otherwise dimensions that are about to be removed still could contain changes.
