+++ Neos.ContentRepository.Core/Classes/Feature/Common/WorkspaceConstraintChecks.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace Neos\ContentRepository\Core\Feature\Common;
 
 use Neos\ContentRepository\Core\CommandHandler\CommandHandlingDependencies;
+use Neos\ContentRepository\Core\Feature\DimensionSpaceAdjustment\Exception\InvalidDimensionAdjustmentTargetWorkspace;
 use Neos\ContentRepository\Core\Feature\WorkspaceCreation\Exception\BaseWorkspaceDoesNotExist;
+use Neos\ContentRepository\Core\SharedModel\Exception\WorkspaceContainsPublishableChanges;
 use Neos\ContentRepository\Core\SharedModel\Exception\WorkspaceDoesNotExist;
 use Neos\ContentRepository\Core\SharedModel\Exception\WorkspaceHasNoBaseWorkspaceName;
 use Neos\ContentRepository\Core\SharedModel\Workspace\Workspace;
 use Neos\ContentRepository\Core\SharedModel\Workspace\WorkspaceName;
+use Neos\ContentRepository\Core\SharedModel\Workspace\Workspaces;
 
 trait WorkspaceConstraintChecks
 {
@@ -40,5 +43,33 @@ trait WorkspaceConstraintChecks
             throw BaseWorkspaceDoesNotExist::butWasSupposedTo($workspace->workspaceName);
         }
         return $baseWorkspace;
+    }
+
+    private function requireWorkspaceToBeRootOrRootBasedForDimensionAdjustment(WorkspaceName $workspaceName, CommandHandlingDependencies $commandHandlingDependencies): void
+    {
+        $workspace = $this->requireWorkspace($workspaceName, $commandHandlingDependencies);
+        if (!$workspace->isRootWorkspace()) {
+            $baseWorkspace = $this->requireBaseWorkspace($workspace, $commandHandlingDependencies);
+            if (!$baseWorkspace->isRootWorkspace()) {
+                throw InvalidDimensionAdjustmentTargetWorkspace::becauseWorkspaceMustBeRootOrRootBased($workspace->workspaceName);
+            }
+        }
+    }
+
+    private static function requireNoWorkspaceToHaveChanges(Workspaces $workspaces, ?WorkspaceName $excludedWorkspaceName): void
+    {
+        $conflictingWorkspaceNames = [];
+        foreach ($workspaces as $workspace) {
+            if ($excludedWorkspaceName?->equals($workspace->workspaceName)) {
+                continue;
+            }
+            if ($workspace->hasPublishableChanges()) {
+                $conflictingWorkspaceNames[] = $workspace->workspaceName;
+            }
+        }
+
+        if ($conflictingWorkspaceNames !== []) {
+            throw WorkspaceContainsPublishableChanges::butWasNotSupposedTo(...$conflictingWorkspaceNames);
+        }
     }
 }
